requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
PyJWT==2.8.0
supabase==2.0.2
openai==1.3.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
